base: Lockscreen Date Fonts [1/2]
Good Job @eldainosor. I was confused because of sliceview shit.
Bootleggers (@eldainosor) edit: Update the code to use the mKeyguardSlice instead of mDateView (deprecated due to google switching all the ls date section to slices) and make a function to apply the same typeface to every view on such SliceView

SDB: Call SliceView(DateView) Fonts with separate method. It looks neat and clean

Signed-off-by: Varun Date <date.varun123@gmail.com>
Signed-off-by: Sipun Ku Mahanta <sipunkumar85@gmail.com>